update_naf

+++ cc_naf.c
@@ -1,9 +1,9 @@
 #include "cc_naf.h"
 
 // return naf_ki = k % w
-static inline int cc_naf_mod(cc_bn_digit_t k, size_t w)
+static inline int8_t cc_naf_mod(cc_bn_digit_t k, size_t w)
 {
-    int ki = k & (((cc_bn_digit_t)1 << w) - 1);
+    int8_t ki = k & (((cc_bn_digit_t)1 << w) - 1);
     if (ki > ((cc_bn_digit_t)1 << (w - 1)))
     {
         ki = (ki - ((cc_bn_digit_t)1 << w));
@@ -13,11 +13,11 @@ static inline int cc_naf_mod(cc_bn_digit_t k, size_t w)
 
 // return naf length
 // note: K is changed
-size_t naf(cc_bn_digit_t *K, size_t K_word_len, size_t w, int *naf)
+size_t naf(cc_bn_digit_t *K, size_t K_word_len, size_t w, int8_t *naf)
 {
     size_t i = 0;
 
-    int ki;
+    int8_t ki;
     while (cc_bn_is_zero(K, K_word_len) == false)
     {
         if (K[0] & 1)
